# cmd/api/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Aplicativo que gerencia atividades de um serviço de pedidos em um restaurante. Desde a base de clientes, catálogo de produtos, pedidos e fila de preparo",
        "title": "CS-EVENTS-API",
        "contact": {},
        "version": "1.0"
    },
    "host": "http://localhost:8081",
    "basePath": "/v1/delivery",
    "paths": {
...
        "/customers/data_removal_request": {
            "post": {
                "description": "save customerdataremoval in DB",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "customer"
                ],
                "summary": "create customerdataremoval",
                "parameters": [
                    {
                        "description": "CustomerDataRemovalRequest",
                        "name": "customerdataremoval",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_Pos-Tech-Challenge-48_delivery-order-api_internal_entities.CustomerDataRemovalRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created"
                    },
                    "400": {
                        "description": "invalid customer data removal",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "general error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "github_com_Pos-Tech-Challenge-48_delivery-order-api_internal_entities.CustomerDataRemovalRequest": {
            "type": "object",
            "properties": {
                "address": {
                    "type": "string"
                },
                "created_date_db": {
                    "type": "string"
                },
                "document": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "last_modified_date_db": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "phonenumber": {
                    "type": "string"
                }
            }
        },
...
    }
}